+++ test/addons-napi/test_object/test_object.c
@@ -3,7 +3,7 @@
 void Get(napi_env env, napi_callback_info info) {
   napi_status status;
 
-  int argc;
+  size_t argc;
   status = napi_get_cb_args_length(env, info, &argc);
   if (status != napi_ok) return;
 
@@ -48,7 +48,7 @@ void Get(napi_env env, napi_callback_info info) {
 void Set(napi_env env, napi_callback_info info) {
   napi_status status;
 
-  int argc;
+  size_t argc;
   status = napi_get_cb_args_length(env, info, &argc);
   if (status != napi_ok) return;
 
@@ -96,7 +96,7 @@ void Set(napi_env env, napi_callback_info info) {
 void Has(napi_env env, napi_callback_info info) {
   napi_status status;
 
-  int argc;
+  size_t argc;
   status = napi_get_cb_args_length(env, info, &argc);
   if (status != napi_ok) return;
 
@@ -169,7 +169,7 @@ void New(napi_env env, napi_callback_info info) {
 void Inflate(napi_env env, napi_callback_info info) {
   napi_status status;
 
-  int argc;
+  size_t argc;
   status = napi_get_cb_args_length(env, info, &argc);
   if (status != napi_ok) return;
